change knockdown to paralyze and sort mechanics
